Take a subset of columns and create a new MTX file-set

Take a subset of columns and create a new MTX file-set

rcpp_mmutil_copy_selected_columns(
  mtx_file,
  row_file,
  col_file,
  r_selected,
  output
)

Arguments

mtx_file
data file

row_file
row file

col_file
column file

selected
selected column names

Examples

options(stringsAsFactors=FALSE)
rr <- rgamma(20, 1, 1)
mm <- matrix(rgamma(10 * 2, 1, 1), 10, 2)
src.hdr <- "test_org"
src.files <- mmutilR::rcpp_mmutil_simulate_poisson(mm, rr, src.hdr)
Y <- Matrix::readMM(src.files$mtx)
colnames(Y) <- read.table(src.files$col)$V1
print(Y)

#> 10 x 19 sparse Matrix of class "dgTMatrix"

#>    [[ suppressing 19 column names ‘1’, ‘2’, ‘3’ ... ]]

#>                                               
#>  [1,] 6 10 . 13 1 10 1 8 1 1 2 2 . 1 1 1 5 1 3
#>  [2,] 1  4 .  1 1  5 . . . 1 3 5 . . 2 . 1 3 3
#>  [3,] .  2 .  . 1  1 . 1 . . 1 . . . . 1 . . .
#>  [4,] .  1 .  1 .  1 . 1 . . . . . . . . 1 . .
#>  [5,] 6  9 2 14 1  9 . 3 4 2 3 1 1 . 1 2 3 1 3
#>  [6,] .  1 .  . .  1 . . . . . 1 . . 1 2 1 1 1
#>  [7,] 1  5 .  7 .  6 1 1 1 . . 1 . 1 2 5 4 3 3
#>  [8,] 2  1 .  1 .  1 . . . . . 2 1 . 1 . . . 1
#>  [9,] .  1 .  2 .  3 . 1 . 1 . 3 . 1 3 4 1 2 4
#> [10,] .  . .  . .  . . . 1 1 2 1 . 1 2 1 4 . 4

sub.cols <- sort(read.table(src.files$col)$V1[sample(20,3)])
print(sub.cols)

#> [1]  7 18

tgt.hdr <- "test_sub"
tgt.files <- mmutilR::rcpp_mmutil_copy_selected_columns(
                         src.files$mtx,
                         src.files$row,
                         src.files$col,
                         sub.cols, tgt.hdr)
Y <- Matrix::readMM(tgt.files$mtx)
colnames(Y) <- read.table(tgt.files$col)$V1
print(Y)

#> 10 x 2 sparse Matrix of class "dgTMatrix"
#>       7 18
#>  [1,] 1  5
#>  [2,] .  1
#>  [3,] .  .
#>  [4,] .  1
#>  [5,] .  3
#>  [6,] .  1
#>  [7,] 1  4
#>  [8,] .  .
#>  [9,] .  1
#> [10,] .  4

unlink(list.files(pattern = src.hdr))
unlink(list.files(pattern = tgt.hdr))