Take a subset of rows and create a new MTX file-set

For the new mtx file, empty columns with only zero elements will be removed.

rcpp_mmutil_copy_selected_rows(
  mtx_file,
  row_file,
  col_file,
  r_selected,
  output
)

Arguments

mtx_file

data file

row_file

row file

col_file

column file

output

output

selected

selected row names

Value

a list of file names: output.mtx,rows,cols.gz

Examples

options(stringsAsFactors=FALSE)
rr <- rgamma(20, 1, 1)
mm <- matrix(rgamma(10 * 2, 1, 1), 10, 2)
src.hdr <- "test_org"
src.files <- mmutilR::rcpp_mmutil_simulate_poisson(mm, rr, src.hdr)
Y <- Matrix::readMM(src.files$mtx)
rownames(Y) <- read.table(src.files$row)$V1
print(Y)
#> 10 x 16 sparse Matrix of class "dgTMatrix"
#>                                     
#> r01 1 . . 2 . . . .  1 . 1 1 . . 1 1
#> r02 2 . . 1 . 1 . .  . 1 . . . . . 1
#> r03 1 . . . . . . .  4 1 1 . . . . 1
#> r04 2 . . 3 1 . 1 2 12 2 1 . 1 . . 2
#> r05 . . . . . . 1 3  5 1 1 . . . . .
#> r06 1 . . . . . . 1  2 1 . . . 2 . .
#> r07 1 . 1 . . 1 . .  . 1 3 . . . . .
#> r08 5 2 1 3 1 . 1 1  8 1 3 . 2 1 . 3
#> r09 3 2 1 2 1 . . .  3 . . . . . . .
#> r10 . . . . . . . .  8 1 2 2 4 . . .
sub.rows <- sort(read.table(src.files$row)$V1[sample(10,3)])
print(sub.rows)
#> [1] "r02" "r03" "r04"
tgt.hdr <- "test_sub"
tgt.files <- mmutilR::rcpp_mmutil_copy_selected_rows(
               src.files$mtx,
               src.files$row,
               src.files$col,
               sub.rows,
               tgt.hdr)
Y <- Matrix::readMM(tgt.files$mtx)
colnames(Y) <- read.table(tgt.files$col)$V1
rownames(Y) <- read.table(tgt.files$row)$V1
print(Y)
#> 3 x 11 sparse Matrix of class "dgTMatrix"
#>    [[ suppressing 11 column names ‘1’, ‘4’, ‘5’ ... ]]
#>                           
#> r02 2 1 . 1 . .  . 1 . . 1
#> r03 1 . . . . .  4 1 1 . 1
#> r04 2 3 1 . 1 2 12 2 1 1 2
unlink(list.files(pattern = src.hdr))
unlink(list.files(pattern = tgt.hdr))